+++ clients/java/cldf-java/src/main/java/io/cldf/models/Manifest.java
@@ -66,6 +66,7 @@ public class Manifest {
     private Integer sessionsCount;
     private Integer locationsCount;
     private Integer routesCount;
+    private Integer sectorsCount;
     private Integer tagsCount;
     private Integer mediaCount;
   }